+++ Expanded Cheat Menu/Code/Console.lua
@@ -358,7 +358,7 @@ function ChoGGi_Funcs.Console.BuildExamineMenu()
 	AddSubmenu("EntityData", nil, "GetAllEntities", "EntityStates", "EntitySurfaces", "HexOutlineShapes", "HexInteriorShapes", "HexOutlineByHash", "HexBuildShapes", "HexBuildShapesInversed", "HexPeripheralShapes", "HexCombinedShapes", "PrefabMarkers")
 	AddSubmenu("g_Classes", nil, "ClassTemplates.Building", "BuildingTemplates", "Attaches", "FXRules", "FXLists")
 	AddSubmenu("g_CObjectFuncs", nil, "hr", "pf", "terrain", "UIL", "DTM", "lpeg", "srp", "camera", "camera3p", "cameraMax", "cameraRTS", "string", "table", "package", "debug", "lfs")
-	AddSubmenu("StoryBits", "Story Bits", "StoryBitCategories", "StoryBitTriggersCombo", "g_StoryBitActive", "g_StoryBitStates", "g_StoryBitCategoryStates")
+	AddSubmenu("StoryBits", "Story Bits", "g_StoryBitsLog", "StoryBitCategories", "StoryBitTriggersCombo", "g_StoryBitActive", "g_StoryBitCategoryStates", "g_StoryBitStates")
 	AddSubmenu("Cities", nil, "GameMaps", "UICity", "MainCity", "UIColony", "UIColony.city_labels.labels", "UIColony.tech_status", "ResupplyItemDefinitions", "BuildMenuPrerequisiteOverrides", "BuildingTechRequirements", "g_ApplicantPool", "g_CurrentMissionParams", "UICity.MapSectors", "RivalAIs", "TaskRequesters", "LRManagerInstance")
 	AddSubmenu("Mods", nil, "ModsLoaded", "ModsList")