+++ Mods ChoGGi/Wide Area Forestation/metadata.lua
@@ -12,9 +12,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2072656358",
 	"pops_any_uuid", "542a9ecf-a195-48ea-8f52-6e716ea4f1be",
 	"lua_revision", 1007000, -- Picard
-	"version", 5,
+	"version", 6,
 	"version_major", 0,
-	"version_minor", 5,
+	"version_minor", 6,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -30,10 +30,11 @@ Adds a button to change all plants to use grid size of selected plant.
 
 
 Mod Options:
-Plant Interval: Plant vegetation interval in hours.
 Remove Power: Remove the requirement for electricity.
 Max Size: Set the max size of the forestation area (can crash when set too high, save before raising).
+Max Size Workaround: 0: Disable workaround, 1: Only show selection hex for selected plant, 2: Never show selection hex.
 
 The game has a hard limit on objects, and selecting a plant makes a lot of objects (each hex is counted as one), so depending on how large your colony is...
 ]],
 })
+--~ Plant Interval: Plant vegetation interval in hours.