+++ Expanded Cheat Menu/Changelog.md
@@ -2,6 +2,7 @@
 
 ## v19.4 (Unreleased)
 ### Added:
+- ECM.Debug.Storybits>Toggle Story Bit Log: Show Storybit info in console log, and extra info in g_StoryBitsLog.
 - ECM.Game.Interface.Toggle Use All Loading Screens: Some DLC replaces loading screens with their own, enable to use all of them.
 - ECM.Debug.Toggles>Show Console Errors: Show extra debug errors in console.
  - This was already part of ECM and enabled by default, it's turned off by default now.
@@ -14,6 +15,9 @@
 
 ### Fixed:
 - When cloning barrel dome at non-default angle grass texture gets stuck on ground.
+
+### Removed:
+- Toggle UI transparency checkbox from console log window.
 
 ## v19.3 (17 Nov 2024)
 ### Changed: